SDOF_Classical_Solution
This code is used to solve second order differential equations of motion using sum of homogenous and particular solution in which two constants of integration are found by complementary solution and are evaluated from prior knowledge of initial conditions. 

1. Eigen Value Analysis - Eigen values and mode shapes
2. Orthogonality of Modes
3. Undamped Free Vibration - SDOF
4. Underdamped Vibration - SDOF
5. Damped Vibration - SDOF

Using this code one can plot the dynamic free vibration response of SDOF system.